src/shot-processor/pyproject.toml
[tool.poetry]
name = "shot-processor"
version = "0.1.0"
description = ""
authors = ["arnoudx6 <arnoud.stolk@ziggo.nl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
configargparse = "^1.7"
pyyaml = "^6.0.1"
types-pyyaml = "^6.0.12.12"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
